2048 메뉴 및 game 파일 추가
