+++ pythonBasic/game/hangman.py
@@ -1,20 +1,4 @@
-# f(x) = 3x + 5 ?
 import random
-
-
-
-def tmpFunction(x):
-    return 3 * x + 5
-
-
-print(tmpFunction(5))
-def menuprint():
-    print("=========GAME=========")
-    print("1. 행맨")
-    print("2. 업다운")
-    print("0, 종료")
-    print("=========GAME=========")
-
 def getRandomWord():
     words = ["hang", "pretty", "apple", "ant", "water", "MCdonalds", "fluent", "voca", "galaxy"]
     return words[random.randrange(0, len(words))]
@@ -70,38 +54,4 @@ def runHangMan():
 
 
     # for문으로 맞는 글자 바꾸기
-def runUpDown():
-    answer = random.randrange(0, 10)
-    chance = 3
-    while chance > 0:
-        user_input = int(input("값을 입력하세요 >>"))
-
-        if user_input == answer:
-            print("정답입니다!")
-            break
-        else:
-            chance = chance - 1
-            if user_input > answer:
-                print("down")
-            else:
-                print("up")
-
-userinput = -1
-
-while userinput != 0:
-    menuprint()
-    userinput = int(input("SELECT MENU ::: "))
-
-    if userinput == 1:
-        runHangMan()
-    elif userinput == 2:
-        runUpDown()
-
-
-
-# 1. 모든 정답을 맞췄을때 게임이 끝나지 않음
-# -> 맞추면 alive  출력해주고 그만하기 (break문을 사용)
-
-# 2. 내가 맞춘 정답들이 어디에 위치해있는지 알수없음
-# -> s _ _ s _ _ _ 출력
-# printCorrectWords() 함수를 선언(optional)해서 그 안에서 입력되었던 맞는 항목을 위치에 맞게 출력
+runHangMan()
